Sorting is now 4x faster with numeric nodes and 3x faster with string nodes.

- **10k random integer string TreeNode**  
  - **Before:** 33 seconds  
  - **After:** 8 seconds

- **String_000.wz/Consume.img** (26k+ properties) in GMS v253  
  - **Before:** 323 seconds  
  - **After:** 95 seconds

- **QuestData_000.wz** (21k+ images) in GMS v253  
  - **Before:** 127 seconds  
  - **After:** 44 seconds

[The `Sort()` method is automatically called when assigning a new `IComparer`](https://referencesource.microsoft.com/#System.Windows.Forms/winforms/Managed/System/WinForms/TreeView.cs,1288).