@lastbattle 2024-09-14
Interesting, also wasnt aware that the int.TryParse and  Convert.ToInt32 is being called twice previously. 
Tested! Thanks.